Refactor of HTTPResponse and Router
