+++ src/com/scottyplunkett/server/Date.java
@@ -1,0 +1,16 @@
+package com.scottyplunkett.server;
+
+import java.text.SimpleDateFormat;
+import java.util.Calendar;
+import java.util.Locale;
+import java.util.TimeZone;
+
+public class Date {
+    static String getDate() {
+        Calendar calendar = Calendar.getInstance();
+        String format = "EEE, dd MMM yyyy HH:mm:ss z";
+        SimpleDateFormat dateFormat = new SimpleDateFormat(format, Locale.US);
+        dateFormat.setTimeZone(TimeZone.getTimeZone("GMT"));
+        return dateFormat.format(calendar.getTime());
+    }
+}
